+++ test/req/adapter_test.exs
@@ -3,6 +3,23 @@ defmodule Req.AdapterTest do
 
   @adapter Req.Case.adapter()
 
+  defmodule StreamResponseStep do
+    def stream(req, acc, fun, state, next) do
+      fun = fn
+        {:headers, _headers} = event, resp, acc, state ->
+          resp = Req.Response.put_header(resp, "x-stream-step", "true")
+          request = Req.Request.put_private(resp.request, :stream_step, true)
+          resp = %{resp | request: request}
+          fun.(event, resp, acc, state)
+
+        event, resp, acc, state ->
+          fun.(event, resp, acc, state)
+      end
+
+      next.(req, acc, fun, state)
+    end
+  end
+
   describe "run" do
     @tag :transport
     test ":inet6" do
@@ -19,18 +36,20 @@ defmodule Req.AdapterTest do
           [:inet6, ip: {0, 0, 0, 0, 0, 0, 0, 1}]
         )
 
-      resp = Req.request!(adapter: adapter_fun(), url: ipv4_url)
+      resp = Req.stream!(adapter: adapter_fun(), url: ipv4_url)
       assert resp.status == 200
       assert resp.body == "ok"
-      resp = Req.request!(adapter: adapter_fun(), url: ipv4_url, inet6: true)
+
+      resp = Req.stream!(adapter: adapter_fun(), url: ipv4_url, inet6: true)
       assert resp.status == 200
       assert resp.body == "ok"
-      resp = Req.request!(adapter: adapter_fun(), url: ipv6_url, inet6: true)
+
+      resp = Req.stream!(adapter: adapter_fun(), url: ipv6_url, inet6: true)
       assert resp.status == 200
       assert resp.body == "ok"
 
       ipv6_url = %{ipv6_url | host: "::1"}
-      resp = Req.request!(adapter: adapter_fun(), url: ipv6_url)
+      resp = Req.stream!(adapter: adapter_fun(), url: ipv6_url)
       assert resp.status == 200
       assert resp.body == "ok"
     end
@@ -49,7 +68,11 @@ defmodule Req.AdapterTest do
       )
 
       resp =
-        Req.request!(adapter: adapter_fun(), url: "http://localhost", unix_socket: socket_path)
+        Req.stream!(
+          adapter: adapter_fun(),
+          url: "http://localhost",
+          unix_socket: socket_path
+        )
 
       assert resp.status == 200
       assert resp.body == "ok"
@@ -79,8 +102,10 @@ defmodule Req.AdapterTest do
         )
 
       assert_receive :accept_queue_filled
-      {:error, err} = Req.request(req)
+      {:error, err, resp} = Req.stream(req)
       assert err == %Req.TransportError{reason: :timeout}
+      assert resp.status == nil
+      assert resp.body == ""
     end
 
     @tag :capture_log
@@ -108,10 +133,10 @@ defmodule Req.AdapterTest do
 
       if @adapter == :httpc do
         assert_raise ArgumentError, "httpc adapter does not support HTTP/2", fn ->
-          Req.request!(req)
+          Req.stream!(req)
         end
       else
-        resp = Req.request!(req)
+        resp = Req.stream!(req)
         assert resp.status == 200
         assert resp.body == "ok"
       end
@@ -132,7 +157,7 @@ defmodule Req.AdapterTest do
         )
 
       assert_raise File.Error, ~r/could not read file "bad.pem"/, fn ->
-        Req.request!(req)
+        Req.stream!(req)
       end
     end
 
@@ -145,15 +170,20 @@ defmodule Req.AdapterTest do
         end)
 
       req = Req.new(adapter: adapter_fun(), url: url, retry: false)
-      {:error, err} = Req.request(req)
+
+      {:error, err, resp} = Req.stream(req)
       assert err == %Req.HTTPError{protocol: :http1, reason: :invalid_status_line}
+      assert resp.status == nil
+      assert resp.body == ""
     end
 
     @tag :transport
     test "Req.TransportError" do
       req = Req.new(adapter: adapter_fun(), url: "http://localhost:9999", retry: false)
-      {:error, err} = Req.request(req)
+      {:error, err, resp} = Req.stream(req)
       assert err == %Req.TransportError{reason: :econnrefused}
+      assert resp.status == nil
+      assert resp.body == ""
     end
 
     @tag :transport
@@ -161,8 +191,10 @@ defmodule Req.AdapterTest do
       req =
         Req.new(adapter: adapter_fun(), url: "http://localhost:9999", inet6: true, retry: false)
 
-      {:error, err} = Req.request(req)
+      {:error, err, resp} = Req.stream(req)
       assert err == %Req.TransportError{reason: :econnrefused}
+      assert resp.status == nil
+      assert resp.body == ""
     end
 
     @tag :transport
@@ -188,8 +220,10 @@ defmodule Req.AdapterTest do
         end)
 
       req = Req.new(adapter: adapter_fun(), url: url, receive_timeout: 50, retry: false)
-      {:error, err} = Req.request(req)
+      {:error, err, resp} = Req.stream(req)
       assert err == %Req.TransportError{reason: :timeout}
+      assert resp.status == nil
+      assert resp.body == ""
       assert_received :ping
     end
 
@@ -209,69 +243,11 @@ defmodule Req.AdapterTest do
         end)
 
       req = Req.new(adapter: adapter_fun(), url: url, request_timeout: 50, retry: false)
-      {:error, err} = Req.request(req)
+      {:error, err, resp} = Req.stream(req)
       assert err == %Req.TransportError{reason: :timeout}
-      assert_receive :ping
-    end
-
-    test "body: req_body_fun succeeded" do
-      %{req: req} =
-        serve(fn conn ->
-          assert {:ok, "foobar", conn} = Plug.Conn.read_body(conn)
-          Plug.Conn.send_resp(conn, 200, "ok")
-        end)
-
-      req_body_fun = fn
-        %Req.Request{private: %{phase: :bar}} = request ->
-          request = Req.Request.put_private(request, :phase, :done)
-          {:data, "bar", request}
-
-        %Req.Request{private: %{phase: :done}} = request ->
-          {:done, request}
-
-        %Req.Request{} = request ->
-          request = Req.Request.put_private(request, :phase, :bar)
-          {:data, "foo", request}
-      end
-
-      {req, resp} = Req.run!(req, method: :post, body: req_body_fun)
-      assert req.private[:phase] == :done
       assert resp.status == 200
-      assert resp.body == "ok"
-    end
-
-    test "body: req_body_fun halted" do
-      %{req: req} =
-        serve(fn conn ->
-          assert {:ok, "", conn} = Plug.Conn.read_body(conn)
-          Plug.Conn.send_resp(conn, 200, "ok")
-        end)
-
-      req_body_fun = fn
-        %Req.Request{} = request ->
-          request = Req.Request.put_private(request, :phase, :halted)
-          {:halt, request}
-      end
-
-      {req, resp} = Req.run!(req, method: :post, body: req_body_fun)
-      assert req.private[:phase] == :halted
-      assert resp.status == nil
       assert resp.body == ""
-    end
-
-    test "body: req_body_fun errored" do
-      %{req: req} =
-        serve(fn conn ->
-          Plug.Conn.send_resp(conn, 200, "ok")
-        end)
-
-      req_body_fun = fn %Req.Request{} -> :oops end
-
-      assert_raise RuntimeError,
-                   "expected req_body_fun to return {:data, chunk, request}, {:done, request}, or {:halt, request}, got: :oops",
-                   fn ->
-                     Req.post!(req, body: req_body_fun)
-                   end
+      assert_receive :ping
     end
 
     @tag :transport
@@ -390,6 +366,25 @@ defmodule Req.AdapterTest do
       end)
     end
 
+    test "into: fun with empty body" do
+      %{req: req} =
+        serve(fn conn ->
+          Plug.Conn.send_resp(conn, 200, "")
+        end)
+
+      {resp, _stderr} =
+        ExUnit.CaptureIO.with_io(:stderr, fn ->
+          Req.request!(req,
+            into: fn {:data, _data}, _acc ->
+              flunk("into: fun called for empty body")
+            end
+          )
+        end)
+
+      assert resp.status == 200
+      assert resp.body == ""
+    end
+
     @tag :transport
     test "into: collectable" do
       %{url: url} =
@@ -417,7 +412,7 @@ defmodule Req.AdapterTest do
         end)
 
       resp =
-        Req.get!(
+        Req.request!(
           adapter: adapter_fun(),
           url: url,
           into: []
@@ -443,7 +438,7 @@ defmodule Req.AdapterTest do
           Req.Test.json(%{conn | status: 404}, %{error: "not found"})
         end)
 
-      resp = Req.get!(req, into: :not_a_collectable)
+      resp = Req.request!(req, into: :not_a_collectable)
 
       assert resp.status == 404
       assert resp.body == %{"error" => "not found"}
@@ -452,7 +447,7 @@ defmodule Req.AdapterTest do
     @tag :transport
     test "into: collectable handle error" do
       {:error, err} =
-        Req.get(
+        Req.request(
           adapter: adapter_fun(),
           url: "http://localhost:9999",
           retry: false,
@@ -479,12 +474,24 @@ defmodule Req.AdapterTest do
           []
         end
 
-      resp = Req.get!(req, connect_options: connect_options, into: :self)
+      resp = Req.request!(req, connect_options: connect_options, into: :self)
       assert resp.status == 200
       assert {:ok, [data: "foo"]} = Req.parse_message(resp, assert_receive(_))
       assert {:ok, [data: "bar"]} = Req.parse_message(resp, assert_receive(_))
       assert {:ok, [:done]} = Req.parse_message(resp, assert_receive(_))
       assert :unknown = Req.parse_message(resp, :other)
+      refute_receive _
+    end
+
+    test "into: :self with empty body" do
+      %{req: req} =
+        serve(fn conn ->
+          Plug.Conn.send_resp(conn, 200, "")
+        end)
+
+      resp = Req.request!(req, into: :self)
+      assert resp.status == 200
+      assert {:ok, [:done]} = Req.parse_message(resp, assert_receive(_))
       refute_receive _
     end
 
@@ -497,7 +504,7 @@ defmodule Req.AdapterTest do
           conn
         end)
 
-      resp = Req.get!(req, into: :self)
+      resp = Req.request!(req, into: :self)
       assert resp.status == 200
       assert :ok = Req.cancel_async_response(resp)
     end
@@ -516,7 +523,7 @@ defmodule Req.AdapterTest do
 
       req = Req.merge(req, url: %{url | path: "/redirect"}, into: :self)
 
-      resp = Req.get!(req)
+      resp = Req.request!(req)
       assert resp.status == 200
       assert Enum.to_list(resp.body) == ["ok"]
     end
@@ -528,7 +535,8 @@ defmodule Req.AdapterTest do
         end)
 
       send(self(), :other)
-      resp = Req.get!(req, into: :self)
+      resp = Req.request!(req, into: :self)
+      assert resp.status == 200
       assert Enum.to_list(resp.body) == ["ok"]
       assert_received :other
     end
@@ -542,7 +550,7 @@ defmodule Req.AdapterTest do
         end)
 
       {:error, err} =
-        Req.get(
+        Req.request(
           adapter: adapter_fun(),
           url: url,
           into: :self,